requirements.txt
Django>=5.1,<6.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
